add photo_view
