+++ happypigs_app/pubspec.yaml
@@ -28,6 +28,7 @@ dependencies:
   path_provider:
   carousel_slider:
   photo_manager:
+  photo_view:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
